% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article.R
\name{get_article}
\alias{get_article}
\title{Scrape meta-data from a single article hosted on Scielo}
\usage{
get_article(url)
}
\arguments{
\item{url}{a character vector with the link of the article hosted on Scielo to be scrapped.}
}
\value{
The function returns an object of class \code{Scielo, data.frame} with the following variables:

\itemize{
  \item author: Author name.
  \item title: Article title.
  \item year: Year of publication.
  \item journal: Journal name.
  \item volume: Volume.
  \item number: Number.
  \item first_page: Article's first page.
  \item last_page: Article's last page
  \item abstratc: Article's abstract.
  \item keywords: Article's keywords.
  \item doi: DOI.
  \item n_authors: Number of authors.
  \item n_pages: Number of pages.
}
}
\description{
\code{get_article()} scrapes meta-data information from an article hosted on Scielo.
}
\details{
This functions scrapes several meta-data information, such as author's names, article title, year of publication, edition and number of pages.
}
\examples{
\dontrun{
article <- get_article(url = "http://www.scielo.br/scielo.php?
script=sci_arttext&pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
summary(article)
}
}

