% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_id.R
\name{get_article_id}
\alias{get_article_id}
\title{Get the ID of a scientific article hosted on Scielo}
\usage{
get_article_id(url)
}
\arguments{
\item{url}{a character vector with the URL of an article hosted on Scielo.}
}
\value{
The function returns a \code{character} vector with the article ID.
}
\description{
\code{get_article_id()} extracts the ID of an article's URL
}
\examples{
\donttest{
id <- get_article_id(url = "http://www.scielo.br/scielo.php?script=sci_arttext&
pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
}
}
