% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_references.R
\name{get_article_references}
\alias{get_article_references}
\title{Scrape bibliographic references from a single article hosted on Scielo}
\usage{
get_article_references(x)
}
\arguments{
\item{x}{a character vector with the link or the id of the article hosted on Scielo
to be scrapped.}
}
\value{
The function a \code{tibble} with the following variables:

\itemize{
  \item references: an article's bibliographic reference (\code{character}).
  \item doi: article's Digital Object Identifier (DOI).
}
}
\description{
\code{get_article_references()} scrapes bibliographic references information
from an article hosted on Scielo.
}
\examples{
\dontrun{
refs <- get_article_references(x = "http://www.scielo.br/scielo.php?
script=sci_arttext&pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
}
}
