% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articleMetaData.R
\name{get_articleMetaData}
\alias{get_articleMetaData}
\title{Scrape meta-data from a single article hosted on Scielo}
\usage{
get_articleMetaData(url)
}
\arguments{
\item{url}{a character vector with the link of the article hosted on Scielo to be scrapped.}
}
\value{
The function returns an object of class \code{Scielo, list} with the following variables:

\itemize{
  \item authors: A data.frame with authors names and last names.
  \item institutions: A list about the author's institutions.
  \item title: A data.frame with the article title in all available languages.
  \item abstract: A data.frame with the article's abstract in all available languages.
  \item keywords: A data.frame with the article's keywords in all available languages.
  \item year: Year of publication.
  \item journal: Journal name.
  \item volume: Journal Volume.
  \item number: Journal Number.
  \item first_page: Article's first page in the journal edition.
  \item last_page: Article's last page in the journal edition.
  \item scielo_id: Article's identification number in the Scielo Database.
  \item doi: DOI.
  \item n_authors: Number of authors.
  \item n_pages: Number of pages.
  \item n_refs: Number of references.
}
}
\description{
\code{get_articleMetaData()} scrapes meta-data information from an article hosted on Scielo.
}
\details{
This functions scrapes several meta-data information, such as author's names, article title, year of publication, edition and number of pages.
}
\examples{
\dontrun{
article <- get_articleMetaData(url = "http://www.scielo.br/scielo.php?
script=sci_arttext&pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
summary(article)
}
}
