% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articleText.R
\name{get_articleText}
\alias{get_articleText}
\title{Scrape the complete text from a single article hosted on Scielo}
\usage{
get_articleText(url)
}
\arguments{
\item{url}{a character vector with the link of the article hosted on Scielo to be scrapped.}
}
\value{
The function returns an object of class \code{Scielo, list} with the following variables:

\itemize{
  \item text: A single-valued character vector containing the entire text.
}
}
\description{
\code{get_articleText()} scrapes the text from an article hosted on Scielo.
}
\details{
[...].
}
\examples{
\dontrun{
article <- get_articleText(url = "http://www.scielo.br/scielo.php?
script=sci_arttext&pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
}
}
