% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_journal.R
\name{get_id_journal}
\alias{get_id_journal}
\title{Get the ID of journal hosted on Scielo}
\usage{
get_id_journal(url)
}
\arguments{
\item{url}{a character vector with the url of the journal hosted on Scielo.}
}
\value{
The function returns an character vector with the journal numerical ID.
}
\description{
\code{get_id_journal()} extracts the numerical ID (pid) from a journal's url.
}
\examples{
\dontrun{
id <- get_id_journal(url = "http://www.scielo.br/scielo.php?
script=sci_serial&pid=1981-3821&lng=en&nrm=iso")
}
}
