% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_id.R
\name{get_journal_id}
\alias{get_journal_id}
\title{Get the ID of a journal hosted on Scielo}
\usage{
get_journal_id(url)
}
\arguments{
\item{url}{a character vector with the URL of a journal hosted on Scielo.}
}
\value{
The function returns a \code{character} vector with the journal ID.
}
\description{
\code{get_journal_id()} extracts the numerical ID (pid) from a journal's URL.
}
\examples{
\donttest{
id <- get_journal_id(url = "http://www.scielo.br/scielo.php?
script=sci_serial&pid=1981-3821&lng=en&nrm=iso")
}
}
