% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journalMetaList.R
\name{get_journalMetaList}
\alias{get_journalMetaList}
\title{Scrape meta-data from all the articles of a journal hosted on Scielo}
\usage{
get_journalMetaList(id_journal)
}
\arguments{
\item{id_journal}{a character vector with the ID of the journal hosted on Scielo (the \code{get_id_journal} function can be used to find the journal ID from its URL).}
}
\value{
The function returns an object of class \code{Scielo, list} with the following variables:

\itemize{
  \item authors: A data.frame with authors names and last names.
  \item institutions: A list about the author's institutions.
  \item title: A data.frame with the article title in all available languages.
  \item abstract: A data.frame with the article's abstract in all available languages.
  \item keywords: A data.frame with the article's keywords in all available languages.
  \item year: Year of publication.
  \item journal: Journal name.
  \item volume: Journal Volume.
  \item number: Journal Number.
  \item first_page: Article's first page in the journal edition.
  \item last_page: Article's last page in the journal edition.
  \item scielo_id: Article's identification number in the Scielo Database.
  \item doi: DOI.
  \item n_authors: Number of authors.
  \item n_pages: Number of pages.
  \item n_refs: Number of references.
}
}
\description{
\code{get_journal()} scrapes meta-data information from all the articles of a journal hosted on Scielo.
}
\details{
This functions scrapes several meta-data information, such as author's names, articles' titles, year of publication, edition and number of pages, that can be summarized with specific \code{summary} method.
}
\examples{
\dontrun{
df <- get_journalMetaList(id_journal = "1981-3821")
summary(df)
}
}
