% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscielo.R
\docType{package}
\name{rscielo-package}
\alias{rscielo}
\alias{rscielo-package}
\title{rscielo: A Scraper for Scientific Journals Hosted on Scielo}
\description{
rscielo provides functions to easily scrape bibliometric information from
scientific journals and articles hosted on the Scientific Electronic Library
Online Platform (Scielo.br). The retrieved data includes a journal's details
and citation counts; article's contents, footnotes, bibliographic references;
and several other common information used in bibliometric studies.
The package also offers functions to quickly summarize the scrapped data.
}
\details{
To learn more about rscielo, check the package documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meirelesff/rscielo}
  \item Report bugs at \url{https://github.com/meirelesff/rscielo/issues}
}

}
\author{
Fernando Meireles [author], Denisson Silva [author],
Rogerio Barbosa [author]
}
