% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_fnotes.R
\name{get_article_fnotes}
\alias{get_article_fnotes}
\title{Scrape foot notes from a single article hosted on Scielo}
\usage{
get_article_fnotes(url)
}
\arguments{
\item{url}{a character vector with the link of the article hosted on Scielo to be scrapped.}
}
\value{
The function returns an object of class \code{Scielo, data.frame} with the following variables:

\itemize{
  \item foot note: foot note
  \item doi: DOI.
}
}
\description{
\code{get_article_fnotes()} scrapes foot note information from an article hosted on Scielo.
}
\examples{
\dontrun{
get_article_fnotes <- get_article_fnotes(url = "http://www.scielo.br/scielo.php?
script=sci_arttext&pid=S1981-38212016000200201&lng=en&nrm=iso&tlng=en")
summary(get_article_fnotes)
}
}
