% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal.R
\name{get_journal}
\alias{get_journal}
\title{Scrape meta-data from all the articles of a journal hosted on Scielo}
\usage{
get_journal(id_journal)
}
\arguments{
\item{id_journal}{a character vector with the ID of the journal hosted on Scielo (the \code{get_id_journal} function can be used to find the journal ID from its url).}
}
\value{
The function returns an object of class \code{Scielo, data.frame} with the following variables:

\itemize{
  \item DATA_GERACAO: Generation date of the file (when the data was collected).
}
}
\description{
\code{get_journal()} scrapes meta-data information of a journal hosted on Scielo.
}
\details{
This functions scrapes several meta-data information, such as author's names, articles' titles, year of publication, edition and number of pages, that can be summarized with specific \code{summary} method.
}
\examples{
\dontrun{
df <- get_journal(id_journal = "1981-3821")
summary(df)
}
}

