% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal.R
\name{get_journal}
\alias{get_journal}
\title{Scrape meta-data from all the articles of a journal hosted on Scielo}
\usage{
get_journal(id_journal)
}
\arguments{
\item{id_journal}{a character vector with the ID of the journal hosted on Scielo (the \code{get_id_journal} function can be used to find the journal ID from its URL).}
}
\value{
The function returns an object of class \code{Scielo, data.frame} with the following variables:

\itemize{
  \item author: Author name.
  \item title: Article title.
  \item year: Year of publication.
  \item journal: Journal name.
  \item volume: Volume.
  \item number: Number.
  \item first_page: Article's first page.
  \item last_page: Article's last page
  \item abstratc: Article's abstract.
  \item keywords: Article's keywords.
  \item doi: DOI.
  \item n_authors: Number of authors.
  \item n_pages: Number of pages.
}
}
\description{
\code{get_journal()} scrapes meta-data information from all the articles of a journal hosted on Scielo.
}
\details{
This functions scrapes several meta-data information, such as author's names, articles' titles, year of publication, edition and number of pages, that can be summarized with specific \code{summary} method.
}
\examples{
\dontrun{
df <- get_journal(id_journal = "1981-3821")
summary(df)
}
}

