% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_info.R
\name{get_journal_info}
\alias{get_journal_info}
\title{Scrape the description of a journal hosted on Scielo}
\usage{
get_journal_info(id_journal)
}
\arguments{
\item{id_journal}{a character vector with the ID of the journal hosted on Scielo (the \code{get_id_journal} function can be used to find the journal ID from its URL).}
}
\value{
The function returns a \code{data.frame} with the journal's description.
}
\description{
\code{get_journal_info()} scrapes the description (publisher, issn, and mission) information of a journal hosted on Scielo.
}
\examples{
\dontrun{
journal_info <- get_journal_info(id_journal = "1981-3821")
}
}

