% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_list.R
\name{get_journal_list}
\alias{get_journal_list}
\title{Scrape a list with all the journals hosted on Scielo}
\usage{
get_journal_list()
}
\value{
The function returns a \code{data.frame} with each journal's title, ID, and URL
}
\description{
\code{get_journal_list()} scrapes the title, numerical ID (pid) and URL of all journals hosted on Scielo.
}
\examples{
\dontrun{
journal_list <- get_journal_list()
}
}
